refactor server metrics: consolidate and optimize server info retrieval and statz visualization
